Dodano ciacho i kontrolę dostępu

+++ src/main/java/pw/paint/controller/RecipeController.java
@@ -17,6 +17,7 @@ import pw.paint.DTOs.requests.NewRecipeRequest;
 import pw.paint.DTOs.requests.SearchRequest;
 import pw.paint.exception.RecipeNotFoundException;
 import pw.paint.exception.UserNotFoundException;
+import pw.paint.service.JwtService;
 import pw.paint.service.RecipeService;
 
 import java.net.URI;
@@ -25,12 +26,14 @@ import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
 import java.util.List;
+import java.util.Objects;
 
 @RestController
 @RequiredArgsConstructor
 @RequestMapping("/recipe")
 public class RecipeController {
     private final RecipeService recipeService;
+    private final JwtService jwtService;
 
     @GetMapping("/tags")
     public ResponseEntity<List<String>> getAllTags() {
@@ -71,8 +74,14 @@ public class RecipeController {
 //    }
 
     @PostMapping("/new")
-    public ResponseEntity<Void> createNewRecipe(@RequestBody NewRecipeRequest newRecipeRequest) {
+    public ResponseEntity<Void> createNewRecipe(@RequestBody NewRecipeRequest newRecipeRequest,
+                                                @CookieValue(name = "token", required = false) String token ) {
         try {
+            if(token == null || !newRecipeRequest.getAuthor().equals(jwtService.extractUsername(token))){
+                return ResponseEntity.status(HttpStatus.FORBIDDEN)
+                        .header("Error-message", "Forbidden")
+                        .build();
+            }
             return ResponseEntity.created(new URI("/recipe/" +
                     recipeService.createNewRecipe(newRecipeRequest))).build();
         } catch (UserNotFoundException ex) {
@@ -87,8 +96,14 @@ public class RecipeController {
     }
 
     @PutMapping("/set-img/{id}")
-    public ResponseEntity<Void> setImage(@PathVariable String id, @ModelAttribute MultipartFile image) {
+    public ResponseEntity<Void> setImage(@PathVariable String id, @ModelAttribute MultipartFile image,
+                                         @CookieValue(name = "token", required = false) String token ) {
         try {
+            if(token == null){
+                return ResponseEntity.status(HttpStatus.FORBIDDEN)
+                        .header("Error-message", "Forbidden")
+                        .build();
+            }
             return ResponseEntity.created(new URI("/recipe/" +
                     recipeService.setImage(id, image))).build();
         } catch (RecipeNotFoundException ex) {
@@ -103,9 +118,18 @@ public class RecipeController {
     }
 
     @GetMapping("/{id}")
-    public ResponseEntity<RecipeDto> getRecipeById(@PathVariable String id) {
+    public ResponseEntity<RecipeDto> getRecipeById(@PathVariable String id,
+                                                   @CookieValue(name = "token", required = false) String token ) {
         try {
-            return ResponseEntity.ok(recipeService.getRecipeById(new ObjectId(id)));
+            var recipe = recipeService.getRecipeById(new ObjectId(id));
+            if(!recipe.getStatus()){
+                if(token == null || !Objects.equals(recipe.getAuthor(), jwtService.extractUsername(token))){
+                    return ResponseEntity.status(HttpStatus.FORBIDDEN)
+                            .header("Error-message", "Forbidden")
+                            .build();
+                }
+            }
+            return ResponseEntity.ok(recipe);
         } catch (RecipeNotFoundException ex) {
             return ResponseEntity.status(HttpStatus.NOT_FOUND)
                     .header("Error-Message", ex.getMessage())
@@ -114,8 +138,15 @@ public class RecipeController {
     }
 
     @PutMapping("/change-status/{id}")
-    public ResponseEntity<Void> changeStatus(@PathVariable String id) {
+    public ResponseEntity<Void> changeStatus(@PathVariable String id,
+                                             @CookieValue(name = "token", required = false) String token ) {
         try {
+            if(token == null ||
+                    !recipeService.getRecipeById(new ObjectId(id)).getAuthor().equals(jwtService.extractUsername(token))){
+                return ResponseEntity.status(HttpStatus.FORBIDDEN)
+                        .header("Error-message", "Forbidden")
+                        .build();
+            }
             recipeService.changeStatus(id);
             return ResponseEntity.ok().build();
         } catch (RecipeNotFoundException ex) {
@@ -133,9 +164,15 @@ public class RecipeController {
     }
 
     @PostMapping("/search/private")
-    public ResponseEntity<List<ShortRecipeDto>> searchPrivate(@RequestBody SearchRequest searchRequest){
+    public ResponseEntity<List<ShortRecipeDto>> searchPrivate(@RequestBody SearchRequest searchRequest,
+                                                              @CookieValue(name = "token", required = false) String token ){
 
         //TO DO sprawdzanie uprawnień czy podany autor w request body to user z tokenu
+        if(token == null || !searchRequest.getAuthor().equals(jwtService.extractUsername(token))){
+            return ResponseEntity.status(HttpStatus.FORBIDDEN)
+                    .header("Error-message", "Forbidden")
+                    .build();
+        }
 
         Pageable pageable = PageRequest.of(searchRequest.getPageNumber(),searchRequest.getPageSize());
         return ResponseEntity.ok(recipeService.search(searchRequest.getAuthor(),
@@ -143,8 +180,16 @@ public class RecipeController {
     }
 
     @DeleteMapping("{recipeId}")
-    public ResponseEntity<Void> deleteRecipe(@PathVariable String recipeId){
+    public ResponseEntity<Void> deleteRecipe(@PathVariable String recipeId,
+                                             @CookieValue(name = "token", required = false) String token ){
         try {
+            if(token == null ||
+                    !recipeService.getRecipeById(new ObjectId(recipeId)).getAuthor()
+                    .equals(jwtService.extractUsername(token))){
+                return ResponseEntity.status(HttpStatus.FORBIDDEN)
+                        .header("Error-message", "Forbidden")
+                        .build();
+            }
             recipeService.deleteRecipe(new ObjectId(recipeId));
             return ResponseEntity.ok().build();
         } catch (RecipeNotFoundException ex) {
